refactor: separate math utilities from DOM operations
